Devolving the custom code to repeated generated code
Using a new http replay value to avoid collision from WAF-329
